Made gtest a subproject.

+++ build.ps1
@@ -5,7 +5,8 @@ param (
     [switch]$compile,
     [switch]$help,
     [switch]$test,
-    [switch]$doc
+    [switch]$doc,
+    [switch]$debug
 )
 
 # Function to display help
@@ -18,9 +19,28 @@ function Show-Help {
     Write-Host "  -compile    Compile the project."
     Write-Host "  -test       Compile and run the tests."
     Write-Host "  -doc        Generate Doxygen documentaion."
+    Write-Host "  -debug      Build debug project."
     Write-Host "  -help       Show this help message."
     Write-Host ""
     Write-Host "If no options are provided, the script will perform the default actions: clean, rebuild, and compile."
+}
+
+# Function to get the current build type from Meson configuration
+function Get-BuildType {
+    if (Test-Path -Path builddir/meson-info/intro-buildoptions.json) {
+        $buildOptions = Get-Content builddir/meson-info/intro-buildoptions.json | ConvertFrom-Json
+        foreach ($option in $buildOptions) {
+            if ($option.name -eq "buildtype") {
+                return $option.value
+            }
+        }
+    }
+    return $null
+}
+
+# Function to setup the Meson build directory
+function MesonSetup([string]$buildtype = "release") {
+    meson setup builddir --buildtype=$buildtype
 }
 
 # Function to clean the build directory
@@ -31,17 +51,43 @@ function Clean {
 
 # Function to rebuild the build directory
 function Rebuild {
-    Write-Host "Rebuilding build directory..."
-    if (Test-Path -Path builddir) {
-        Remove-Item -Recurse -Force builddir
+    Write-Host "Checking current build type..."
+    $currentBuildType = Get-BuildType
+    if ($null -eq $currentBuildType) {
+        Write-Host "Build directory does not exist. Setting up a new build directory."
+        if ($debug) {
+            MesonSetup "debug"
+        }
+        else {
+            MesonSetup "release"
+        }
     }
-    meson setup builddir
+    elseif (($debug -and $currentBuildType -ne "debug") -or (-not $debug -and $currentBuildType -eq "debug")) {
+        Write-Host "Current build type is $currentBuildType. Rebuilding for the correct build type..."
+        if (Test-Path -Path builddir) {
+            Remove-Item -Recurse -Force builddir
+        }
+        if ($debug) {
+            MesonSetup "debug"
+        }
+        else {
+            MesonSetup "release"
+        }
+    }
+    else {
+        Write-Host "Current build type is already $currentBuildType. No rebuild needed."
+    }
 }
 
 # Function to compile the project
 function Compile {
     Write-Host "Compiling the project..."
     meson compile -C builddir
+
+    if (-not $?) {
+        Write-Host "Compilation failed. Exiting."
+        exit 1
+    }
 }
 
 # Function to run the tests
@@ -80,7 +126,6 @@ if ($test) {
 
 if ($doc) {
     Write-Doxygen
-    exit
 }
